Welp, adding color is going to get tricky.
Need to handle the fact that ANSI color codes leave the cursor
unaffected. Also, \t probably does magic in the terminal I'll have
to figure out as well. Honestly, we're probalby hosed when it comes
to dealing with \t.

Honestly, it'd probably be easier to just delete the entire line and
rerender it. Though if that takes more then a render frame that could
be bad? Might want to use absolute positioning, which might(?) be
easier.